move Sections enum to dargling.docstring.sections module
