+++ darglint/docstring/base.py
@@ -1,5 +1,4 @@
 from abc import ABC, abstractmethod
-import enum
 from typing import (
     Callable,
     Dict,
@@ -10,18 +9,8 @@ from typing import (
     Iterable,
 )
 
-from ..strictness import Strictness
-
-
-class Sections(enum.Enum):
-    SHORT_DESCRIPTION = 0
-    LONG_DESCRIPTION = 1
-    ARGUMENTS_SECTION = 2
-    RAISES_SECTION = 4
-    YIELDS_SECTION = 6
-    RETURNS_SECTION = 8
-    VARIABLES_SECTION = 10
-    NOQAS = 13
+from .sections import Sections  # noqa: F401
+from ..strictness import Strictness  # noqa: F401
 
 
 class BaseDocstring(ABC):
